// src/orval/type.schemas.ts
export interface UserDto {
  /** @nullable */
  companyName?: string | null;
  createdAt?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  id?: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  picFullName?: string | null;
  /** @nullable */
  profilePictureUrl?: string | null;
  roleName?: RoleName;
  /** @nullable */
  roles?: string[] | null;
  updatedAt?: string;
  /** @nullable */
  username?: string | null;
}
// ...
export type StatesAndTerritoriesOfMalaysia = typeof StatesAndTerritoriesOfMalaysia[keyof typeof StatesAndTerritoriesOfMalaysia];



export const StatesAndTerritoriesOfMalaysia = {
  Default: 'Default',
  Johor: 'Johor',
  Kedah: 'Kedah',
  Kelantan: 'Kelantan',
  Malacca: 'Malacca',
  NegeriSembilan: 'NegeriSembilan',
  Pahang: 'Pahang',
  Penang: 'Penang',
  Perak: 'Perak',
  Perlis: 'Perlis',
  Sabah: 'Sabah',
  Sarawak: 'Sarawak',
  Selangor: 'Selangor',
  Terengganu: 'Terengganu',
  KualaLumpur: 'KualaLumpur',
  Labuan: 'Labuan',
  Putrajaya: 'Putrajaya',
} as const;

export interface SellRentRequestDetailDto {
  /** @nullable */
  hourMeter?: string | null;
  id?: string;
  isRent?: boolean;
  location?: StatesAndTerritoriesOfMalaysia;
  /** @nullable */
  machineMake?: string | null;
  /** @nullable */
  machineType?: string | null;
  /** @nullable */
  modelNo?: string | null;
  regCard?: boolean;
  requestId?: string;
}
// ...
export type RequestType = typeof RequestType[keyof typeof RequestType];



export const RequestType = {
  Default: 'Default',
  Buy: 'Buy',
  Sell: 'Sell',
  Rent: 'Rent',
  RentOut: 'RentOut',
  Insurance: 'Insurance',
  Dealer: 'Dealer',
  Collaborator: 'Collaborator',
  Advertisement: 'Advertisement',
  Enquiry: 'Enquiry',
} as const;

export type RequestStatus = typeof RequestStatus[keyof typeof RequestStatus];



export const RequestStatus = {
  Default: 'Default',
  Pending: 'Pending',
  Appointment: 'Appointment',
  PendingPayment: 'PendingPayment',
  PendingDelivery: 'PendingDelivery',
  Completed: 'Completed',
  Incomplete: 'Incomplete',
} as const;
// ...
export interface RequestLogDto {
  appointment?: AppointmentDto;
  createdAt?: string;
  createdBy?: string;
  createdByUser?: UserDto;
  fromStatus?: RequestStatus;
  id?: string;
  purchase?: PurchaseDto;
  /** @nullable */
  remark?: string | null;
  requestId?: string;
  /** @nullable */
  requestLogMedias?: RequestLogMediaDto[] | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  toStatus?: RequestStatus;
}

export interface RequestLogDtoPagedResult {
  /** @nullable */
  items?: RequestLogDto[] | null;
  itemsFrom?: number;
  itemsTo?: number;
  totalItemsCount?: number;
  totalPages?: number;
}

export interface RequestDto {
  /** @nullable */
  appointments?: AppointmentDto[] | null;
  buyRentRequestDetail?: BuyRentRequestDetailDto;
  collaboratorRequestDetail?: CollaboratorRequestDetailDto;
  createdAt?: string;
  dealerRequestDetail?: DealerRequestDetailDto;
  enquiryDetail?: EnquiryDetailDto;
  id?: string;
  insuranceRequestDetail?: InsuranceRequestDetailDto;
  /** @nullable */
  requestLogs?: RequestLogDto[] | null;
  /** @nullable */
  requestorEmail?: string | null;
  /** @nullable */
  requestorName?: string | null;
  /** @nullable */
  requestorPhoneNumber?: string | null;
  requestStatus?: RequestStatus;
  requestType?: RequestType;
  sellRentRequestDetail?: SellRentRequestDetailDto;
  updatedAt?: string;
  user?: UserDto;
}

export interface RequestDtoPagedResult {
  /** @nullable */
  items?: RequestDto[] | null;
  itemsFrom?: number;
  itemsTo?: number;
  totalItemsCount?: number;
  totalPages?: number;
}

export interface RequestDtoDataResponseDto {
  data?: RequestDto;
  /** @nullable */
  errMessage?: string | null;
  isSuccess?: boolean;
}
// ...
export interface MachineTypeDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  typeIconUrl?: string | null;
}
// ...
export interface MachineMakeDto {
  id?: string;
  /** @nullable */
  logoImageUrl?: string | null;
  /** @nullable */
  name?: string | null;
}
// ...
export type InsuranceType = typeof InsuranceType[keyof typeof InsuranceType];



export const InsuranceType = {
  Default: 'Default',
  HeavyEquipmentAllRiskInsurance: 'HeavyEquipmentAllRiskInsurance',
  EquipmentAllRiskInsurance: 'EquipmentAllRiskInsurance',
  ContractorAllRiskInsurance: 'ContractorAllRiskInsurance',
  WorkmanCompensation: 'WorkmanCompensation',
  PerformanceBond: 'PerformanceBond',
  Others: 'Others',
} as const;

export interface InsuranceRequestDetailDto {
  id?: string;
  insuranceType?: InsuranceType;
  location?: StatesAndTerritoriesOfMalaysia;
  /** @nullable */
  machineMake?: string | null;
  /** @nullable */
  machineType?: string | null;
  /** @nullable */
  modelName?: string | null;
  registrationCard?: boolean;
  /** @nullable */
  registrationNo?: string | null;
  requestId?: string;
  /** @nullable */
  sumInsured?: string | null;
}
// ...
export interface FilterDto {
  /** @nullable */
  machineMakes?: MachineMakeDto[] | null;
  /** @nullable */
  machineTypes?: MachineTypeDto[] | null;
  totalMachineMakes?: number;
  totalMachineTypes?: number;
}
// ...
export interface EnquiryDetailDto {
  /** @nullable */
  enquiryRemarks?: string | null;
  id?: string;
  requestId?: string;
}
// ...
export interface DealerRequestDetailDto {
  /** @nullable */
  dealershipName?: string | null;
  id?: string;
  location?: StatesAndTerritoriesOfMalaysia;
  requestId?: string;
}
// ...
export interface CreateInsuranceRequestCommand {
  insuranceType?: InsuranceType;
  location?: StatesAndTerritoriesOfMalaysia;
  /** @nullable */
  machineMake?: string | null;
  /** @nullable */
  machineType?: string | null;
  /** @nullable */
  modelName?: string | null;
  registrationCard?: boolean;
  /** @nullable */
  registrationNo?: string | null;
  /** @nullable */
  requestorEmail?: string | null;
  /** @nullable */
  requestorName?: string | null;
  /** @nullable */
  requestorPhoneNumber?: string | null;
  /** @nullable */
  sumInsured?: string | null;
}
// ...
export interface CollaboratorRequestDetailDto {
  /** @nullable */
  companyName?: string | null;
  id?: string;
  /** @nullable */
  remarks?: string | null;
  requestId?: string;
}
// ...
export interface BuyRentRequestDetailDto {
  id?: string;
  interestType?: InterestType;
  isRent?: boolean;
  listing?: AdminListingDto;
  listingId?: string;
  requestId?: string;
}
// ...
export interface AppointmentDto {
  address?: AddressDto;
  appointmentDateTime?: string;
  appointmentStatus?: AppointmentStatus;
  appointmentType?: AppointmentType;
  businessAffiliate?: BusinessAffiliateSummaryDto;
  createdAt?: string;
  id?: string;
  request?: RequestSummaryDto;
  updatedAt?: string;
  user?: UserDto;
}